Change totalDays, usedDays type int->double

+++ src/main/java/musai/app/DTO/request/UserLeaveRequestDTO.java
@@ -9,8 +9,8 @@ public class UserLeaveRequestDTO {
 	private Long Id;
     private Long userId;
     private Long leaveTypeId;
-    private Integer totalDays;
-    private Integer usedDays;
+    private Double totalDays;
+    private Double usedDays;
     private LocalDate validFrom;
     private LocalDate validTo;
